v0.3 [Extending and Restructing Configuration, Adding Configuration Commands, Improving Fuzzy Finding]
- Extended the configuration struct and functionality
- Better flow from app configuration to commands
- Commands that augment configuration from the terminal
- Improved the fuzzy finding of the open functionality. It now includes
finding notes based on their tags as well as note title.
- Added a system to break out of the "atoms" only note taking structure.
These are molecules, and they allow you to create a separate folder to
use. In general, the atoms should be the main flow of notes however
their are uses cases for organizing particular notes such as for writing
blog posts or documentation for an application you're going to publish.
In those cases it makes sense to branch them into their own molecule.
- Improved error handling, feedback messages, and validation methods.
- Extended the base ZettelkastenNote struct with new functionality