Add Symlink Command
Symlink command allows fuzzyfinding a note and dropping a symlinked
version of it in the current working directory. This allows a user to
quickly make edits to a commonly used note within a project by simply
adding it to their project working directory for easy repeated access
that links right back to their vault.